requirements.txt
gradio==4.26.0
groq==0.4.2
pandas==2.2.1
scikit-learn==1.4.1.post1
python-dotenv==1.0.1